Moved views, urls, forms, and filters into their new app locations
