+++ apps/books/tests/test_models.py
@@ -8,66 +8,63 @@ from datetime import datetime, timedelta
 
 class BookModelTest(TestCase):
     def setUp(self):
-        self.user = User.objects.create_user(username='testuser', password='12345')
-        self.library = Library.objects.create(name='Test Library', owner=self.user)
+        self.user = User.objects.create_user(username="testuser", password="12345")
+        self.library = Library.objects.create(name="Test Library", owner=self.user)
 
     def test_str(self):
         book = Book.objects.create(
             library=self.library,
-            name='Test Book',
+            name="Test Book",
             quantity=5,
-            author='Author',
-            category='Education',
-            language='English'
+            author="Author",
+            category="Education",
+            language="English",
         )
         self.assertEqual(str(book), "Test Book [5]")
 
     def test_meta_app_label(self):
-        self.assertEqual(Book._meta.app_label, 'library')
+        self.assertEqual(Book._meta.app_label, "library")
 
     def test_category_choices_uppercase(self):
-        self.assertTrue(hasattr(Book, 'CATEGORY_CHOICES'))
+        self.assertTrue(hasattr(Book, "CATEGORY_CHOICES"))
         self.assertIsInstance(Book.CATEGORY_CHOICES, list)
 
     def test_language_choices_uppercase(self):
-        self.assertTrue(hasattr(Book, 'LANGUAGE_CHOICES'))
+        self.assertTrue(hasattr(Book, "LANGUAGE_CHOICES"))
         self.assertIsInstance(Book.LANGUAGE_CHOICES, list)
 
 
 class IssuedBookModelTest(TestCase):
     def setUp(self):
-        self.user = User.objects.create_user(username='student', password='12345')
-        self.library = Library.objects.create(name='Test Library', owner=self.user)
+        self.user = User.objects.create_user(username="student", password="12345")
+        self.library = Library.objects.create(name="Test Library", owner=self.user)
         self.student = StudentExtra.objects.create(
             user=self.user,
             library=self.library,
-            name='Student Name',
-            enrollment='12345'
+            name="Student Name",
+            enrollment="12345",
         )
         self.book = Book.objects.create(
-            library=self.library,
-            name='Test Book',
-            quantity=5,
-            author='Author'
+            library=self.library, name="Test Book", quantity=5, author="Author"
         )
 
     def test_str(self):
         issued = IssuedBook.objects.create(
             student=self.student,
             book=self.book,
-            enrollment='12345',
-            book_name='Test Book'
+            enrollment="12345",
+            book_name="Test Book",
         )
         self.assertEqual(str(issued), "12345 - Test Book")
 
     def test_save_auto_fill(self):
         issued = IssuedBook(student=self.student, book=self.book)
         issued.save()
-        self.assertEqual(issued.enrollment, '12345')
-        self.assertEqual(issued.book_name, 'Test Book')
+        self.assertEqual(issued.enrollment, "12345")
+        self.assertEqual(issued.book_name, "Test Book")
 
     def test_meta_app_label(self):
-        self.assertEqual(IssuedBook._meta.app_label, 'library')
+        self.assertEqual(IssuedBook._meta.app_label, "library")
 
 
 class GetExpiryTest(TestCase):
